requirements.txt
fastapi
uvicorn[standard]
python-dotenv
langchain
langchain-huggingface
langchain-communities
chromadb
openai
